requirements.txt
# Secure Tool Runner - Dependencies
# Core
httpx>=0.27.0          # Async HTTP client for Ollama API
pydantic>=2.0.0        # Data validation and settings
pyyaml>=6.0.0          # YAML config loading
rich>=13.0.0           # Beautiful terminal output
typer>=0.12.0          # CLI interface

# 1Password Integration
# Uses 1Password CLI (op) - install separately:
# brew install 1password-cli
